changed how image url is assigned for post, added utils class
